requirements.txt
django~=5.1.7
djangorestframework~=3.15.2
django~=filter-25.1
djangorestframework-simplejwt~=5.5.0
drf-yasg~=1.21.10
Pillow~=11.1.0
django-jazzmin~=ww3.0.1